🚑♻️🚚🎨 Refactor GPT integration to OCR API.
